pyproject.toml
[project]
name = "alchemist-shell"
version = "0.1.0"
description = ""
authors = [
    {name = "Elkana Maina",email = "elisoft.engineer@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy>=2.0.0",
    "typer>=0.9.0",
    "ipython>=8.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0"
]

[tool.poetry]
packages = [{include = "alchemist_shell", from = "src"}]

[project.scripts]
alchemist = "alchemist_shell.cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
